Introduce a deleteprefix option
Now it is possible to delete all the links pointing to directories which
share a common prefix. For example `goat deleteprefix $HOME/Documents`
will delete all the links which point to $HOME/Documents or its
subdirectory.

This feature is useful when you clean up goat after finishing a project
when you want to delete all the links related to that project.