Avoid broken ln on mingw
Use `cmd /c mklink /D` instead of ln on MinGW systems. This uses a
helper batch script to get the arguments passed correctly through
MinGW's path-mangling.

On MinGW, ln is implemented as a copy function. which defeats the intent
of the command. Unfortunately, mklink requires admin privileges on
Windows, but at least that will work in more cases than the previous
version.